+++ src/templates/components/charts/index.ts
@@ -2,3 +2,4 @@ export { BarChart } from "./BarChart";
 export { Histogram } from "./Histogram";
 export { LineChart } from "./LineChart";
 export { MixedChart } from "./MixedChart";
+export { PieChart } from "./PieChart";